Revert "MEPA-326 I2C read/write mapping with MEPA for M10G"
This reverts commit 51290bec3d56dbec148ef0347e74080b9ce207e0.

+++ mepa/vtss/include/vtss_phy_10g_api.h
@@ -3013,8 +3013,8 @@ vtss_rc vtss_phy_10g_i2c_reset(const vtss_inst_t                inst,
 
 vtss_rc vtss_phy_10g_i2c_read(const vtss_inst_t                inst,
                               const vtss_port_no_t             port_no,
-                              const u8                         addr,
-                              u8                               *value);
+                              const u16                        addr,
+                              u16                              *value);
 /**
  * \brief  Write to i2c device
  *
@@ -3028,8 +3028,8 @@ vtss_rc vtss_phy_10g_i2c_read(const vtss_inst_t                inst,
 
 vtss_rc vtss_phy_10g_i2c_write(const vtss_inst_t                inst,
                                const vtss_port_no_t             port_no,
-                               const u8                         addr,
-                               const u8                         *value);
+                               const u16                        addr,
+                               const u16                        *value);
 
 /** \brief 10G Phy I2C Master Interface for SFP Module Configuration */
 typedef struct {
